Fix AreaJump for steam, WndHandle adjustments
-Fix AreaJump not working in steam version.
There is a bunch of different
executables for this game online, the one steam downloads is packed with
steamStub and has 1.0.0.0 version string (which works for the most part
the only thing that failed was area jump location), there are also some
executables with version string 1.1.0.0 which look like they have
completely different offsets(?). Maybe calculate exe md5 hashes or
something to warn the user/ dump info into the logfile.
-Grab Window handle from DX9 Device Creation Parameters, might fix gui
not responding to input for some people(?).